fix: copy and clean edge case

+++ utils/clean.js
@@ -2,6 +2,8 @@ import fs from 'fs'
 
 export default (dirs) => {
     dirs.forEach((dir) => {
-        fs.rmSync(dir, { recursive: true })
+        if(fs.existsSync(dir)) {
+            fs.rmSync(dir, { recursive: true })
+        }
     })
 }
