+++ utils/copy.js
@@ -1,7 +1,9 @@
-import fs from 'fs-extra'
+import fs from "fs-extra"
 
 export default (dirs) => {
     for (const dir of dirs) {
-        fs.copySync(dir.src, dir.dest)
+        if (fs.existsSync(dir)) {
+            fs.copySync(dir.src, dir.dest)
+        }
     }
 }
